Move player mixin
